updates controllers authorization

+++ app/controllers/users_controller.rb
@@ -1,7 +1,5 @@
 class UsersController < ApplicationController
 
-  before_action :authenticate_user, except: [:create]
-
   def create
     user = User.new(
       name: params[:name],
@@ -19,27 +17,39 @@ class UsersController < ApplicationController
 
   def show
     user = User.find(params[:id])
-    render json: user
+    if current_user.id == user.id
+      render json: user
+    else
+      render json: { errors: "Unauthorized" }, status: :unauthorized
+    end
   end
 
   def update
     user = User.find(params[:id])
-    user.name = params[:name] || user.name
-    user.email = params[:email] || user.email
-    user.image_url = params[:image_url] || user.image_url
-    if user.save
-      render json: { message: "User successfully updated" },
-      status: :ok
+    if current_user.id == user.id
+      user.name = params[:name] || user.name
+      user.email = params[:email] || user.email
+      user.image_url = params[:image_url] || user.image_url
+      if user.save
+        render json: { message: "User successfully updated" },
+        status: :ok
+      else
+        render json: { errors: user.errors.full_messages },
+        status: :bad_request
+      end
     else
-      render json: { errors: user.errors.full_messages },
-      status: :bad_request
+      render json: { errors: "Unauthorized" }, status: :unauthorized
     end
   end
 
   def destroy
     user = User.find(params[:id])
-    user.destroy
-    render json: { message: "User successfully deleted"}
+    if current_user == user
+      user.destroy
+      render json: { message: "User successfully deleted"}
+    else
+      render json: { errors: "Unauthorized" }, status: :unauthorized
+    end
   end
 
 end
